requirements.txt
# Core dependencies for getmail.py
pyyaml>=6.0
imapclient>=2.3.0
html2text>=2020.1.16
python-dotenv>=0.19.0
PySocks>=1.7.1

# Dependencies for readmail_qwen.py
openai>=1.0.0
